🚚MOVE: 엔티티 디렉터리 변경하기(#4)
domain 디렉터리 안에 dto, denttiy 디렉터리로 구분하기 위해 엔티티 클래스들을 옮겼다. 생각보다 디렉터리 혹은 클래스가 많이 생성될 것으로 예상되기 때문에 생각날 때 옮겼다.
