+++ src/main/java/com/dedication/force/domain/entity/Article.java
@@ -1,21 +1,25 @@
-package com.dedication.force.domain;
+package com.dedication.force.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 import java.time.ZonedDateTime;
+import java.util.Set;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class Comment {
+public class Article {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "comment_id", nullable = false, updatable = false)
+    @Column(name = "article_id", nullable = false, updatable = false)
     private Long id;
 
-    @Column(length = 200, nullable = false)
+    @Column(length = 50, nullable = false)
+    private String title;
+
+    @Column(length = 3000, nullable = false)
     private String content;
 
     @Column(nullable = false)
@@ -27,4 +31,7 @@ public class Comment {
     @JoinColumn(name = "member_id")
     @ManyToOne(fetch = FetchType.LAZY)
     private Member member;
+
+    @OneToMany(mappedBy = "article", cascade = CascadeType.ALL, orphanRemoval = true)
+    private Set<ArticleCategory> articleCategories;
 }
